Feat: add hand prefab with animation and connect each secen with object
