+++ Assets/AnimateHandOnInput.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 99b4a737903b82a43bb95804a2149147
+guid: 23ef43102b78d074ab1005da97469381
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
