--- Assets/Oculus Hands Physics/Animations.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5312cf64a580ce242aadae4fa36765df
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
