--- Assets/Oculus Hands Physics/Animations/Left Hand Animator.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0284ecac7be7e2d409126f62358d744b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
