--- Assets/Oculus Hands Physics/Prefabs/Custom Left Hand Model.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7811abe38615a86438f679bcf79f5b92
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
