--- Assets/Oculus Hands Physics/Prefabs/Custom Right Hand Model.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d2121c8d819597349ac6f3e6043fa1f3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
