--- Assets/Oculus Hands Physics/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fe37bd8dfb4a37b499eab19f16f57fdb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
