+++ Assets/Oculus Hands.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: ddc38db09f9124241b4b4e9b2b9831eb
-PrefabImporter:
+guid: fc418d5ed5159f24b9212a11bb97db1d
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
