+++ Assets/Oculus Hands/Animations.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 6c2509c76b281144cb0099fc45f17a09
-PrefabImporter:
+guid: e840c6d7927a8794895d56387046aea7
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
