+++ Assets/Oculus Hands/Animations/Left Hand Animator.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 5bea1e892b5c61f4e83c76c6ebd566bd
-folderAsset: yes
-DefaultImporter:
+guid: 4b5efa0c69d901449a38989140500c0d
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
