+++ Assets/Oculus Hands/Animations/Right Hand Animator.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 93756a6a02237c0418736f2a31b54230
-folderAsset: yes
-DefaultImporter:
+guid: 3338dde1ae80a3b43826deca5c238356
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
