+++ Assets/Oculus Hands/Animations/l_hand_default_anim.fbx
