+++ Assets/Oculus Hands/Animations/l_hand_fist_anim.fbx.meta
