+++ Assets/Oculus Hands/Animations/l_hand_pinch_anim.fbx.meta
