+++ Assets/Oculus Hands/Animations/r_hand_default_anim.fbx
