+++ Assets/Oculus Hands/Animations/r_hand_fist_anim.fbx
