+++ Assets/Oculus Hands/Animations/r_hand_pinch_anim.fbx
