+++ Assets/Oculus Hands/Materials/Hands_solid.mat.meta
