+++ Assets/Oculus Hands/Models/l_hand_skeletal_lowres.fbx.meta
