+++ Assets/Oculus Hands/Models/r_hand_skeletal_lowres.fbx.meta
