+++ Assets/Oculus Hands/Prefabs.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 77459436eeb3aeb48871d657bcb5fd7f
-PrefabImporter:
+guid: 176a87d49f06d4340bf832ac8e2f5df5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
