+++ Assets/Oculus Hands/Prefabs/Left Hand Model.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3c99d0ba506360a43b8cea08ef805d25
+guid: 65847589f926b734d9eaed6e25f0046b
 PrefabImporter:
   externalObjects: {}
   userData:
