+++ Assets/Oculus Hands/Prefabs/Right Hand Model.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 011fededf8deb2f479afbcb453d11410
-NativeFormatImporter:
+guid: 1fd3a7c43c8190f4d8c167ebde694364
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
