+++ Assets/Oculus Hands/Scripts.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: e6fdd4b9480500e4ea28cad030e5ddc2
-NativeFormatImporter:
+guid: 38af7a4c7e82ba14fb147fbe5a55e8a3
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
